+++ CHANGELOG.md
@@ -12,6 +12,10 @@
 
 - Add the `FeedbackButton` component that shows the Feedback Widget ([#4378](https://github.com/getsentry/sentry-react-native/pull/4378))
 
+### Fixes
+
+- Considers the `SENTRY_DISABLE_AUTO_UPLOAD` and `SENTRY_DISABLE_NATIVE_DEBUG_UPLOAD` environment variables in the configuration of the Sentry Android Gradle Plugin for Expo plugin ([#4583](https://github.com/getsentry/sentry-react-native/pull/4583))
+
 ### Dependencies
 
 - Bump Bundler Plugins from v3.2.0 to v3.2.1 ([#4585](https://github.com/getsentry/sentry-react-native/pull/4585))